+++ internal/provider/datasource_escalation_chains.go
@@ -58,6 +58,7 @@ func (d *EscalationChainsDataSource) Schema(_ context.Context, _ datasource.Sche
 									"fallback_to_all":  schema.BoolAttribute{Computed: true},
 									"user_id":          schema.StringAttribute{Computed: true},
 									"webhook_url":      schema.StringAttribute{Computed: true},
+									"headers":          schema.MapAttribute{Computed: true, Sensitive: true, ElementType: types.StringType},
 									"delay_minutes":    schema.Int64Attribute{Computed: true},
 									"tracker_type":     schema.StringAttribute{Computed: true},
 									"url":              schema.StringAttribute{Computed: true},